add 'RandomizedQueryOptimizerBase' as an abstract base class
