+++ src/main/java/se/liu/ida/hefquin/engine/queryproc/impl/optimizer/randomized/IterativeImprovementBasedQueryOptimizer.java
@@ -2,67 +2,49 @@ package se.liu.ida.hefquin.engine.queryproc.impl.optimizer.randomized;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
-import java.util.Set;
-import se.liu.ida.hefquin.engine.queryplan.LogicalPlan;
 import se.liu.ida.hefquin.engine.queryplan.PhysicalPlan;
 import se.liu.ida.hefquin.engine.queryproc.QueryOptimizationException;
-import se.liu.ida.hefquin.engine.queryproc.QueryOptimizer;
-import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.CostModel;
 import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.QueryOptimizationContext;
-import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.evolutionaryAlgorithm.RuleApplicationsOfPlans;
-import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.rewriting.RuleApplication;
-import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.rewriting.RuleInstances;
 import se.liu.ida.hefquin.engine.queryproc.impl.optimizer.utils.CostEstimationUtils;
 import se.liu.ida.hefquin.engine.utils.Pair;
 
-public class IterativeImprovementBasedQueryOptimizer implements QueryOptimizer {
-	
-	protected final QueryOptimizationContext context;
+public class IterativeImprovementBasedQueryOptimizer extends RandomizedQueryOptimizerBase
+{
 	protected final StoppingConditionForIterativeImprovement condition;
-	protected final Random rng  = new Random();
-	protected final RuleApplicationsOfPlans rules = new RuleApplicationsOfPlans( new RuleInstances() );
-	
-	public IterativeImprovementBasedQueryOptimizer (final QueryOptimizationContext ctxt, final StoppingConditionForIterativeImprovement x) {
-		assert ctxt != null;
+
+	public IterativeImprovementBasedQueryOptimizer( final QueryOptimizationContext context,
+	                                                final StoppingConditionForIterativeImprovement x ) {
+		super(context);
+
 		assert x != null;
-		context = ctxt;
 		condition = x;
 	}
-	
-	
-	@Override
-	public PhysicalPlan optimize( final LogicalPlan initialPlan ) throws QueryOptimizationException {
-		return optimize( context.getLogicalToPhysicalPlanConverter().convert(initialPlan,false) );
-	}
 
+	@Override
 	public PhysicalPlan optimize( final PhysicalPlan initialPlan ) throws QueryOptimizationException {
 		// The best plan we have found so far. As we have only found one plan, it is the best one so far.
 		PhysicalPlan bestPlan = initialPlan;
-		
-		CostModel costModel = context.getCostModel(); //Created as a value and stored rather than just using the getCostModel, because the same model is used later too.
 
-		final Double initialCost = CostEstimationUtils.getEstimates(costModel, initialPlan)[0];
-		
+		final Double initialCost = CostEstimationUtils.getEstimates( context.getCostModel(), initialPlan )[0];
+
 		// generation = number of times the outer loop has run. Has to be declared here since it will increment each outer loop.
 		int generation = 0;
-		
+
 		// bestCost = best possible cost out of everything that the algorithm has found. Has to be declared here so that we have something to compare to in all the iterations of the outer loop without losing track of it between loops.
 		Double bestCost = initialCost;
-		
-		while(!condition.readyToStop(generation)) { // Currently only handles generation number as a stopping condition!
-			
+
+		while ( !condition.readyToStop(generation) ) { // Currently only handles generation number as a stopping condition!
 			// The randomized plan generator is to be used here. As a temporary measure, the initial plan is used.
 			PhysicalPlan currentPlan = initialPlan; // This variable will hold the plan which is currently being worked on.
 			Double currentCost = initialCost; // The cost of the current plan. For it to be a local minimum, none of its neighbours can have a lower cost
 			boolean improvementFound = false;
-			
+
 			do {
 				final List<PhysicalPlan> neighbours = getNeighbours(currentPlan);
-				final Double[] neighbourCosts = CostEstimationUtils.getEstimates(costModel, neighbours);
+				final Double[] neighbourCosts = CostEstimationUtils.getEstimates( context.getCostModel(), neighbours );
 				final List<Pair<PhysicalPlan,Double>> betterPlans = new ArrayList<>();
 				improvementFound = false;
-				
+
 				// Using a for-loop in order to have the index for which neighbouring plan to pick.
 				for (int x = 0; x < neighbourCosts.length; x++) {
 					if(neighbourCosts[x] < currentCost) {
@@ -71,7 +53,7 @@ public class IterativeImprovementBasedQueryOptimizer implements QueryOptimizer {
 						betterPlans.add(betterPlan); // I haven't figured out how to do this properly for Java yet, but it will be solved by the PhysicalPlanWithCost anyway.
 					}
 				}
-				
+
 				if(improvementFound) { // if we have found at least one possible improvement, we want to make one of them into our new current plan.
 					final Pair<PhysicalPlan,Double> newPlan = betterPlans.get(rng.nextInt(betterPlans.size())); // Get a random object.
 					currentPlan = newPlan.object1;
@@ -79,27 +61,16 @@ public class IterativeImprovementBasedQueryOptimizer implements QueryOptimizer {
 				}
 			}
 			while(improvementFound);
-			
+
 			if(currentCost < bestCost) {
 				bestCost = currentCost;
 				bestPlan = currentPlan;
 			}
-			
+
 			generation++;
 		}
-		
+
 		return bestPlan;
 	}
-	
-	protected List<PhysicalPlan> getNeighbours(final PhysicalPlan initialPlan) {
-		final List<PhysicalPlan> resultList = new ArrayList<PhysicalPlan>();
-		
-		final Set<RuleApplication> ruleApplications = rules.getRuleApplications(initialPlan);
-		
-		for ( final RuleApplication ra : ruleApplications ) {
-			resultList.add( ra.getResultingPlan() );
-		}
-		
-		return resultList;
-	}
+
 }
